Moving share route to separate layout
