+++ apps/endatix-hub/app/(main)/error.tsx
