+++ apps/endatix-hub/app/(main)/favicon.ico
