+++ apps/endatix-hub/app/(public)/share/[formId]/layout.tsx
@@ -11,8 +11,13 @@ export default function Layout({
   children: React.ReactNode;
 }>) {
   return (
-    <div>
-      {children}
-    </div>
+    <html lang="en">
+      <body>
+        <header></header>
+        <main className="grid flex-1 items-start gap-4 p-4 sm:px-6 sm:py-0 md:gap-8">
+          {children}
+        </main>
+      </body>
+    </html>
   );
 }
