+++ apps/endatix-hub/components/survey-creator.tsx
@@ -5,9 +5,9 @@ import { startTransition } from "react";
 import { useRouter } from "next/navigation";
 import { ICreatorOptions } from "survey-creator-core";
 import { SurveyCreatorComponent, SurveyCreator } from "survey-creator-react";
-import { updateFormDefinitionJsonAction } from "../app/forms/[formId]/update-form-definition-json.action"
-import { updateFormNameAction } from "@/app/forms/[formId]/update-form-name.action";
-import { updateFormStatusAction } from "@/app/forms/[formId]/update-form-status.action";
+import { updateFormDefinitionJsonAction } from "../app/(main)/forms/[formId]/update-form-definition-json.action"
+import { updateFormNameAction } from "@/app/(main)/forms/[formId]/update-form-name.action";
+import { updateFormStatusAction } from "@/app/(main)/forms/[formId]/update-form-status.action";
 import "survey-core/defaultV2.css";
 import "survey-creator-core/survey-creator-core.css";
